Simple Sender + Receiver Application with 2 debuggable components.
To debug a component, add <instance>.debug = "True"; in your
configuration to debug that component.

Each component also has a deactivation framework in the source code
that the user can code to turn off the GDB hooks. For example,
component A can only be debugged if there is 5V on Pin10.

“Approved for Public Release, Distribution Unlimited”

Change-Id: I58dbcf2e338a1440d58a360a00e7b0dbab3a8a08